requirements.txt
# Production
py-cord
pyyaml

# Dev
pylint
pytest

